Move TrayIcon to separate file
